# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Diagnostics": {
        "additionalProperties": false,
        "description": "What the extractors could not see around a node: unresolved references whose\nsource or file is the node, and extraction issues of a file.",
        "properties": {
          "extraction_issues": {
            "items": {
              "$ref": "#/components/schemas/ExtractionIssue"
            },
            "title": "Extraction Issues",
            "type": "array"
          },
          "node": {
            "$ref": "#/components/schemas/NodeBrief"
          },
          "unresolved_references": {
            "items": {
              "$ref": "#/components/schemas/UnresolvedReference"
            },
            "title": "Unresolved References",
            "type": "array"
          }
        },
        "required": [
          "node",
          "unresolved_references",
          "extraction_issues"
        ],
        "title": "Diagnostics",
        "type": "object"
      },
...
      "ExtractionIssue": {
        "additionalProperties": false,
        "description": "A problem an extractor hit: parse errors, unsupported syntax, rejected paths.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "file_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Key"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "severity": {
            "$ref": "#/components/schemas/IssueSeverity"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceSpan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "severity",
          "code",
          "message",
          "provenance"
        ],
        "title": "ExtractionIssue",
        "type": "object"
      },
...
      "IssueSeverity": {
        "enum": [
          "error",
          "warning",
          "info"
        ],
        "title": "IssueSeverity",
        "type": "string"
      },
      "NodeBrief": {
        "additionalProperties": false,
        "description": "Enough to name and link to a node.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "repository",
              "directory",
              "file",
              "type",
              "function",
              "endpoint",
              "package",
              "external_package"
            ],
            "title": "SnapshotNodeKind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "kind",
          "name"
        ],
        "title": "NodeBrief",
        "type": "object"
      },
...
      "Page_NodeSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NodeSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque; pass as `cursor` to get the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[NodeSummary]",
        "type": "object"
      },
...
      "Page_SearchHit_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque; pass as `cursor` to get the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[SearchHit]",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "An error, as RFC 9457 problem details (``application/problem+json``).",
        "properties": {
          "code": {
            "description": "Stable machine-readable code, e.g. node_not_found",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-parameter details of a validation_error",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URN identifying the problem type: urn:xraymyrepo:problem:<code>",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "UnresolvedReference": {
        "additionalProperties": false,
        "description": "A reference found in source whose target could not be determined.",
        "properties": {
          "candidates": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Candidates",
            "type": "array"
          },
          "end_col": {
            "minimum": 0.0,
            "title": "End Col",
            "type": "integer"
          },
          "end_line": {
            "minimum": 1.0,
            "title": "End Line",
            "type": "integer"
          },
          "file_key": {
            "title": "File Key",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/UnresolvedReason"
          },
          "ref_kind": {
            "$ref": "#/components/schemas/ReferenceKind"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "start_col": {
            "minimum": 0.0,
            "title": "Start Col",
            "type": "integer"
          },
          "start_line": {
            "minimum": 1.0,
            "title": "Start Line",
            "type": "integer"
          }
        },
        "required": [
          "start_line",
          "start_col",
          "end_line",
          "end_col",
          "source_key",
          "file_key",
          "ref_kind",
          "raw_text",
          "reason",
          "provenance"
        ],
        "title": "UnresolvedReference",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Read-only access to persisted Codebase Intelligence Model snapshots.",
    "title": "XRayMyRepo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/snapshots/{snapshot_id}/diagnostics": {
      "get": {
        "operationId": "diagnostics_api_v1_snapshots__snapshot_id__diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "description": "CIM node key",
            "in": "query",
            "name": "node",
            "required": true,
            "schema": {
              "description": "CIM node key",
              "title": "Node",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagnostics"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Snapshot not complete"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameters or key"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "What the extractors could not see around a node",
        "tags": [
          "nodes"
        ]
      }
    },
...
    "/api/v1/snapshots/{snapshot_id}/nodes": {
      "get": {
        "description": "In key (code point) order. Use it for nodes outside the tree: packages,\nexternal packages and endpoints.",
        "operationId": "list_nodes_api_v1_snapshots__snapshot_id__nodes_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "description": "Repeatable filter",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "repository",
                      "directory",
                      "file",
                      "type",
                      "function",
                      "endpoint",
                      "package",
                      "external_package"
                    ],
                    "title": "SnapshotNodeKind",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable filter",
              "title": "Kind"
            }
          },
          {
            "description": "Key prefix, e.g. `src/app/`",
            "in": "query",
            "name": "prefix",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Key prefix, e.g. `src/app/`",
              "title": "Prefix"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "description": "Page size",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_cursor` of the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_cursor` of the previous page",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NodeSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid cursor"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Snapshot not complete"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameters or key"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "List nodes by kind and/or key prefix",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/nodes/by-key": {
      "get": {
        "operationId": "get_node_api_v1_snapshots__snapshot_id__nodes_by_key_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "description": "CIM node key",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "description": "CIM node key",
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Snapshot not complete"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameters or key"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "Inspect one node",
        "tags": [
          "nodes"
        ]
      }
    },
...
    "/api/v1/snapshots/{snapshot_id}/search": {
      "get": {
        "description": "Case-insensitive substring match. Exact names first, then name prefixes, then\nother name matches, then nodes whose key (path) contains the text.",
        "operationId": "search_api_v1_snapshots__snapshot_id__search_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "description": "Text to find",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Text to find",
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Repeatable filter",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "repository",
                      "directory",
                      "file",
                      "type",
                      "function",
                      "endpoint",
                      "package",
                      "external_package"
                    ],
                    "title": "SnapshotNodeKind",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable filter",
              "title": "Kind"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "description": "Page size",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_cursor` of the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_cursor` of the previous page",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SearchHit_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid cursor"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Snapshot not complete"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameters or key"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "Find nodes by name, key or path",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/v1/snapshots/{snapshot_id}/tree/children": {
      "get": {
        "description": "Directories, then files (by key); inside a file, types and functions in source\norder, then endpoints. Each child carries ``child_count`` so it can show an\nexpander without another request. Packages are not in the tree; list them with\n``/nodes?kind=package``.",
        "operationId": "tree_children_api_v1_snapshots__snapshot_id__tree_children_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "description": "Key of the node to expand",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "default": "/",
              "description": "Key of the node to expand",
              "title": "Parent",
              "type": "string"
            }
          },
          {
            "description": "Repeatable filter",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "repository",
                      "directory",
                      "file",
                      "type",
                      "function",
                      "endpoint",
                      "package",
                      "external_package"
                    ],
                    "title": "SnapshotNodeKind",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable filter",
              "title": "Kind"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "description": "Page size",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_cursor` of the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_cursor` of the previous page",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NodeSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid cursor"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Snapshot not complete"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameters or key"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "Expand one node of the declaration tree",
        "tags": [
          "nodes"
        ]
      }
    },
...
  }
}